# Fixed ArrayPoolBufferWriter<T> repeated new[] allocations