--- a/UnitTests/UnitTests.HighPerformance.Shared/Buffers/Test_ArrayPoolBufferWriter{T}.cs
+++ b/UnitTests/UnitTests.HighPerformance.Shared/Buffers/Test_ArrayPoolBufferWriter{T}.cs
@@ -6,6 +6,7 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Linq;
+using System.Reflection;
 using Microsoft.Toolkit.HighPerformance.Buffers;
 using Microsoft.Toolkit.HighPerformance.Extensions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -17,6 +18,37 @@ namespace UnitTests.HighPerformance.Buffers
     [SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1649", Justification = "Test class for generic type")]
     public class Test_ArrayPoolBufferWriterOfT
     {
+        [TestCategory("ArrayPoolBufferWriterOfT")]
+        [TestMethod]
+        [DataRow(0, 256)] // 256 is the default initial size for ArrayPoolBufferWriter<T>
+        [DataRow(4, 256)]
+        [DataRow(7, 256)]
+        [DataRow(27, 256)]
+        [DataRow(188, 256)]
+        [DataRow(257, 512)]
+        [DataRow(358, 512)]
+        [DataRow(799, 1024)]
+        [DataRow(1024, 1024)]
+        [DataRow(1025, 2048)]
+        [DataRow((1024 * 1024) - 1, 1024 * 1024)]
+        [DataRow(1024 * 1024, 1024 * 1024)]
+        [DataRow((1024 * 1024) + 1, 2 * 1024 * 1024)]
+        [DataRow(2 * 1024 * 1024, 2 * 1024 * 1024)]
+        [DataRow((2 * 1024 * 1024) + 1, 4 * 1024 * 1024)]
+        [DataRow(3 * 1024 * 1024, 4 * 1024 * 1024)]
+        public void Test_ArrayPoolBufferWriterOfT_BufferSize(int request, int expected)
+        {
+            using var writer = new ArrayPoolBufferWriter<byte>();
+
+            writer.GetSpan(request);
+
+            var arrayFieldInfo = typeof(ArrayPoolBufferWriter<byte>).GetField("array", BindingFlags.Instance | BindingFlags.NonPublic);
+
+            byte[] array = (byte[])arrayFieldInfo!.GetValue(writer);
+
+            Assert.AreEqual(array!.Length, expected);
+        }
+
         [TestCategory("ArrayPoolBufferWriterOfT")]
         [TestMethod]
         public void Test_ArrayPoolBufferWriterOfT_AllocateAndGetMemoryAndSpan()